tensorflow_framework/requirements_tensorflow.txt
# --
# requirements for tensorflow framework

dask
docker
jinja2
librosa
keras
matplotlib
numpy
numba
pandas
tdqm
tensorflow
pyarrow
pyyaml
pydantic


# removed these:
#kaleido
#plotly